added freelancers and method for set hourly and fixed rate

+++ src/main/java/model/Employee.java
@@ -9,11 +9,11 @@ public interface Employee {
      * Метод, устанавливающий фиксированную ставку
      * @param fixedRate
      */
-    void setFixedRate(float fixedRate);
+    void setFixedRate(int fixedRate);
 
     /**
      * Метод возвращает фиксированную ставку
      * @return fixedRate
      */
-    float getFixedRate();
+    int getFixedRate();
 }
